server/requirements.txt
Django==3.2.3
PyJWT==2.1.0
boto3
boto3==1.17.84
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.11.1
djangorestframework-camel-case==1.2.0
djangorestframework==3.12.4
drf-dynamic-fields==0.3.1
drf-writable-nested==0.6.3
psycopg2-binary==2.8.6

pytest
requests
uWSGI
